Fix compilation problems with newer toolchains.
The biggest change is removing the use of readdir_r which is deprecated.
We didn't really need it, because each thred opens the directory itself.
According to the readdir_r man page on Linux, "However, in modern
implementations (including the glibc implementation), concurrent calls to
readdir(3) that specify different directory streams  are  thread-safe."